REFACTOR: refactor team config page. Now works with components
