+++ src/components/visor/teams/config/index.ts
@@ -1,0 +1,7 @@
+export * from "./Caminante";
+export * from "./Structure";
+export * from "./Auxiliar";
+export * from "./Enlace";
+export * from "./Miembros";
+export * from "./GeoArea";
+export * from "./PointType";
